requirements.txt
cassandra-driver
lz4
websocket-client
pusherclient
requests
boto3